third_party/WebKit/LayoutTests/fast/css/attribute-selector-case-insensitive-ascii.html

<!DOCTYPE html>
<title>ASCII case-insensitive attribute selector matching</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="element1" attr-exact="f&#x130;x"></div>
<div id="element2" attr-exact="a&#x308;pple"></div>
<div id="element3" attr-list="apelsin a&#x308;pple citron"></div>
<div id="element4" attr-hyphen="a&#x308;ppel-cider"></div>
<div id="element5" attr-begin="a&#x308;ppelcider"></div>
<div id="element6" attr-end="gifta&#x308;pple"></div>
<div id="element7" attr-contain="glasa&#x308;ppelkorg"></div>
<script>
function matchingId(selector) {
    return (document.querySelector(selector) || { id: '' }).id;
}
test(function() {
    assert_equals(matchingId('div[attr-exact="f\u0130x"]'), 'element1');
    assert_equals(matchingId('div[attr-exact="FIX" i]'), '');
    assert_equals(matchingId('div[attr-exact="A\u0308pple" i]'), 'element2');
    assert_equals(matchingId('div[attr-list~="A\u0308pple" i]'), 'element3');
    assert_equals(matchingId('div[attr-hyphen|="A\u0308ppel" i]'), 'element4');
    assert_equals(matchingId('div[attr-begin^="A\u0308ppel" i]'), 'element5');
    assert_equals(matchingId('div[attr-end$="A\u0308pple" i]'), 'element6');
    assert_equals(matchingId('div[attr-contain*="A\u0308ppel" i]'), 'element7');
});
</script>
